clean up identifiers defined in `Main`
- Loaded packages do not need explicit bindings
- The name `MainInclude` does not need to be visible
- Put Main's eval and include in the module like all other modules and
  hide them explicitly instead
